Intellij code suggestions Delete unused ImagePlusToOMETiff class Removes OIR patch since bf PR has been merged Logging modifications Removes useless multiseries file name checks in ImagePlus to OME Tiff exporter Improves handling of ImagePlus calibration unit finder Fix documentation formatting

+++ src/main/java/ch/epfl/biop/kheops/ometiff/AverageImageScaler.java
@@ -108,7 +108,7 @@ public class AverageImageScaler implements IImageScaler {
 							| ((allshifts[is][offs + 1] & 0xFF) << 16)
 							| ((allshifts[is][offs + 2] & 0xFF) << 8)
 							| (allshifts[is][offs + 3] & 0xFF);
-					v += Float.intBitsToFloat(asInt);
+					v += (int) Float.intBitsToFloat(asInt);
 				}
 				buffer.putFloat(v / fnShifts);
 				offs += bytesPerPixel;
